Common coordinate transformation
