+++ .vscode/settings.json
@@ -1,0 +1,7 @@
+{
+    "files.associations": {
+        "*.sycl": "cpp",
+        "array": "cpp",
+        "type_traits": "cpp"
+    }
+}
