+++ math/cmath/include/algebra/math/cmath.hpp
@@ -8,12 +8,9 @@
 #pragma once
 
 // Impl include(s).
-#include "algebra/math/impl/cmath_cartesian2.hpp"
-#include "algebra/math/impl/cmath_cylindrical2.hpp"
 #include "algebra/math/impl/cmath_getter.hpp"
 #include "algebra/math/impl/cmath_matrix.hpp"
 #include "algebra/math/impl/cmath_operator.hpp"
-#include "algebra/math/impl/cmath_polar2.hpp"
 #include "algebra/math/impl/cmath_transform3.hpp"
 #include "algebra/math/impl/cmath_vector.hpp"
 // Algorithms include(s).
@@ -22,3 +19,8 @@
 #include "algebra/math/algorithms/matrix/inverse/cofactor.hpp"
 #include "algebra/math/algorithms/matrix/inverse/hard_coded.hpp"
 #include "algebra/math/algorithms/utils/algorithm_finder.hpp"
+
+// Common include(s)
+namespace algebra::cmath {
+#include "algebra/coordinate/coordinates.hpp"
+}  // namespace algebra::cmath
