+++ math/common/include/algebra/coordinate/coordinates.hpp
@@ -1,0 +1,14 @@
+/** Algebra plugins library, part of the ACTS project
+ *
+ * (c) 2022 CERN for the benefit of the ACTS project
+ *
+ * Mozilla Public License Version 2.0
+ */
+
+#pragma once
+
+// Project include(s).
+#include "algebra/coordinate/cartesian2.hpp"
+#include "algebra/coordinate/cylindrical2.hpp"
+#include "algebra/coordinate/polar2.hpp"
+#include "algebra/coordinate/radial.hpp"
