+++ math/common/include/algebra/coordinate/polar2.hpp
@@ -1,6 +1,6 @@
 /** Algebra plugins library, part of the ACTS project
  *
- * (c) 2020-2022 CERN for the benefit of the ACTS project
+ * (c) 2022 CERN for the benefit of the ACTS project
  *
  * Mozilla Public License Version 2.0
  */
@@ -8,22 +8,18 @@
 #pragma once
 
 // Project include(s).
-#include "algebra/math/impl/eigen_getter.hpp"
 #include "algebra/qualifiers.hpp"
 
-namespace algebra::eigen::math {
-
-/** Local frame projection into a polar coordinate frame */
+/** Local frame projection into a polar coordinate frame
+ */
 template <typename transform3_t>
-struct cylindrical2 {
+struct polar2 {
 
   /// @name Type definitions for the struct
   /// @{
 
   /// Transformation matching this struct
   using transform3_type = transform3_t;
-  /// Scalar type used by the transform
-  using scalar_type = typename transform3_type::scalar_type;
 
   /// Point in 2D space
   using point2 = typename transform3_type::point2;
@@ -31,18 +27,6 @@ struct cylindrical2 {
   using point3 = typename transform3_type::point3;
 
   /// @}
-
-  /** This method transform from a point from 2D or 3D cartesian frame to a 2D
-   * polar point
-   *
-   * @param v the point in local frame
-   *
-   * @return a local point2
-   */
-  ALGEBRA_HOST_DEVICE inline point2 operator()(const point3 &v) const {
-
-    return {perp(v) * phi(v), v[2]};
-  }
 
   /** This method transform from a point from the global 3D cartesian frame to
    *the local 2D cartesian frame
@@ -57,6 +41,18 @@ struct cylindrical2 {
 
     return operator()(trf.point_to_local(p));
   }
-};  // struct cylindrical2
 
-}  // namespace algebra::eigen::math
+  /** This method transform from a point in 2D cartesian frame to a 2D
+   * polar point */
+  ALGEBRA_HOST_DEVICE inline point2 operator()(const point2 &v) const {
+
+    return {perp(v), phi(v)};
+  }
+  /** This method transform from a point in 3D cartesian frame to a 2D
+   * polar point */
+  ALGEBRA_HOST_DEVICE inline point2 operator()(const point3 &v) const {
+
+    return {perp(v), phi(v)};
+  }
+
+};  // struct polar2
