+++ math/eigen/CMakeLists.txt
@@ -7,11 +7,8 @@
 # Set up the library.
 algebra_add_library( algebra_eigen_math eigen_math
    "include/algebra/math/eigen.hpp"
-   "include/algebra/math/impl/eigen_cartesian2.hpp"
-   "include/algebra/math/impl/eigen_cylindrical2.hpp"
    "include/algebra/math/impl/eigen_getter.hpp"
    "include/algebra/math/impl/eigen_matrix.hpp"
-   "include/algebra/math/impl/eigen_polar2.hpp"
    "include/algebra/math/impl/eigen_transform3.hpp"
    "include/algebra/math/impl/eigen_vector.hpp" )
 target_link_libraries( algebra_eigen_math
