+++ math/eigen/include/algebra/math/eigen.hpp
@@ -8,9 +8,12 @@
 #pragma once
 
 // Project include(s).
-#include "algebra/math/impl/eigen_cylindrical2.hpp"
 #include "algebra/math/impl/eigen_getter.hpp"
 #include "algebra/math/impl/eigen_matrix.hpp"
-#include "algebra/math/impl/eigen_polar2.hpp"
 #include "algebra/math/impl/eigen_transform3.hpp"
 #include "algebra/math/impl/eigen_vector.hpp"
+
+// Common include(s)
+namespace algebra::eigen::math {
+#include "algebra/coordinate/coordinates.hpp"
+}  // namespace algebra::eigen::math
