+++ math/smatrix/CMakeLists.txt
@@ -10,12 +10,9 @@ find_package( ROOT COMPONENTS Core MathCore Smatrix REQUIRED )
 # Set up the library.
 algebra_add_library( algebra_smatrix_math smatrix_math
    "include/algebra/math/smatrix.hpp"
-   "include/algebra/math/impl/smatrix_cartesian2.hpp"
-   "include/algebra/math/impl/smatrix_cylindrical2.hpp"
    "include/algebra/math/impl/smatrix_errorcheck.hpp"
    "include/algebra/math/impl/smatrix_getter.hpp"
    "include/algebra/math/impl/smatrix_matrix.hpp"
-   "include/algebra/math/impl/smatrix_polar2.hpp"
    "include/algebra/math/impl/smatrix_transform3.hpp"
    "include/algebra/math/impl/smatrix_vector.hpp" )
 target_link_libraries( algebra_smatrix_math
